Refactor: Split NeatConfig.ts into focused modules (#1599)
Extract NeatConfig.ts (1,120 lines) into focused modules:
- NeatConfigParsers.ts: Sub-config parsing functions (plateau, stability, etc.)
- NeatConfigValidation.ts: Cross-field validation logic

NeatConfig.ts is now a ~515-line factory that delegates complex
sub-config parsing and validation to the extracted modules.
